@@ -6,7 +6,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -14,6 +14,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   ffi: ^2.0.1
   flutter_rust_bridge: ^1.63.1
+  file_picker: ^5.2.5
 
 dev_dependencies:
   flutter_test:
